+++ CleanArch.Application/Features/Feedbacks/IFeedbackService.cs
@@ -8,12 +8,12 @@ namespace CleanArch.Application.Features.Feedbacks
 {
     public interface IFeedbackService
     {
-        Task<Result<Exception, FeedbackOutputModel>> ApproveAsync(FeedbackApproveModel inputModel, long feedbackId);
+        Task<Result<Exception, FeedbackOutputModel>> ApproveAsync(FeedbackApproveModel inputModel, int feedbackId);
 
         Task<Result<Exception, FeedbackOutputModel>> AddAsync(FeedbackInputModel inputModel);
 
         Task<Result<Exception, IQueryable<FeedbackOutputModel>>> GetAll();
 
-        Task<Result<Exception, FeedbackOutputModel>> GetByIdAsync(long id);
+        Task<Result<Exception, FeedbackOutputModel>> GetByIdAsync(int id);
     }
 }
